decouples monitor from internal and swaps timer to histogram mpMetrics
