+++ dev/io.openliberty.mcp.internal.monitor/src/io/openliberty/mcp/internal/monitor/McpStats.java
@@ -11,6 +11,9 @@ package io.openliberty.mcp.internal.monitor;
 
 import com.ibm.websphere.monitor.meters.StatisticsMeter;
 import com.ibm.websphere.monitor.meters.StatisticsReading;
+
+import io.openliberty.mcp.internal.monitoring.McpStatAttributes;
+
 import com.ibm.websphere.monitor.meters.Counter;
 import com.ibm.websphere.monitor.meters.Meter;
