+++ dev/io.openliberty.microprofile.metrics.5.0.internal/src/io/openliberty/microprofile/metrics50/mcp/MPMetricsMcpMetricsAdapterImpl.java
@@ -44,8 +44,8 @@ import com.ibm.ws.container.service.state.ApplicationStateListener;
 import com.ibm.ws.container.service.state.StateChangeException;
 import com.ibm.ws.runtime.metadata.ComponentMetaData;
 
-import io.openliberty.mcp.internal.monitor.McpStatAttributes;
 import io.openliberty.mcp.internal.monitor.metrics.McpMetricAdapter;
+import io.openliberty.mcp.internal.monitoring.McpStatAttributes;
 import io.openliberty.microprofile.metrics50.SharedMetricRegistries;
 import io.openliberty.microprofile.metrics50.helper.Constants;
