+++ dev/io.openliberty.microprofile.telemetry.2.0.internal/bnd.bnd
@@ -57,6 +57,7 @@ Import-Package: \
 DynamicImport-Package: \
     io.openliberty.http.monitor,\
     io.openliberty.http.monitor.metrics,\
+    io.openliberty.mcp.internal.monitoring,\
     io.openliberty.mcp.internal.monitor,\
     io.openliberty.mcp.internal.monitor.metrics,\
     com.ibm.ws.connectionpool.monitor.metrics
@@ -86,6 +87,7 @@ DynamicImport-Package: \
 	io.opentelemetry.semconv:opentelemetry-semconv;version='1.25.0.alpha';strategy=exact,\
 	com.ibm.ws.kernel.boot.common;version=latest,\
 	io.openliberty.http.monitor,\
+	io.openliberty.mcp.internal;version=latest,\
 	io.openliberty.mcp.internal.monitor,\
 	io.opentelemetry.instrumentation:opentelemetry-resources;version='2.1.0.alpha',\
 	com.ibm.ws.connectionpool.monitor,\
